api/requirements.txt
fastapi>=0.128.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.30.0,<1.0.0
pydantic-settings>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[argon2]>=1.7.0,<2.0.0
argon2-cffi>=23.1.0,<24.0.0
httpx>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
email-validator>=2.0.0,<3.0.0
icalendar>=5.0.11,<6.0.0
alembic>=1.14.0,<2.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
